Move TextLink under Button
Ensure LoadingButton tests render children

fixup! Ensure LoadingButton tests render children
