+++ src/components/Button/components/PlainButton/index.js
@@ -1,0 +1,3 @@
+import PlainButton from './PlainButton';
+
+export default PlainButton;
