+++ src/components/Button/components/RegularButton/index.js
@@ -1,0 +1,3 @@
+import RegularButton from './RegularButton';
+
+export default RegularButton;
