--- src/components/TextLink/index.js
@@ -1,3 +1,0 @@
-import TextLink from './TextLink';
-
-export default TextLink;
